LayoutTests/svg/as-image/svg-canvas-svg-with-feimage-with-link-tainted-expected.txt

CONSOLE MESSAGE: line 1: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
PASS window.ctx.getImageData(0, 0, 1, 1) threw exception SecurityError (DOM Exception 18): The operation is insecure..
